requirements.txt
# Backend
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
psycopg2-binary
pgvector

# AI / Embeddings
pyautogen
sentence-transformers
openai

# Telegram Bot
aiogram>=3.4.0

# Scraper
requests
beautifulsoup4
selenium
webdriver-manager

# Utilities
python-dotenv
numpy
aiofiles